make the suggested changes
