+++ Modules/HR/Config/config.php
@@ -175,5 +175,8 @@ return [
     'hr-followup-email' => [
         'primary' => env('HR_FOLLOWUP_EMAIL_PRIMARY', 'deepak.sharma@coloredcow.in'),
         'secondary' => env('HR_FOLLOWUP_EMAIL_SECONDARY', 'pk@coloredcow.in'),
-    ]
+    ],
+    'url' => [
+        'requisition_url' => env('REQUISITION_URL'),
+    ],
 ];
