+++ Modules/HR/Http/Controllers/EmployeeController.php
@@ -8,10 +8,7 @@ use Modules\HR\Entities\Employee;
 use Illuminate\Routing\Controller;
 use Modules\HR\Entities\HrJobDomain;
 use Modules\HR\Entities\Job;
-use Modules\HR\Entities\JobRequisition;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
-use Illuminate\Support\Facades\Mail;
-use Modules\HR\Emails\SendHiringMail;
 
 class EmployeeController extends Controller
 {
@@ -56,31 +53,13 @@ class EmployeeController extends Controller
     {
         $domainId = $request->domain_id;
         $employees = Employee::where('domain_id', $domainId)->get();
-        $DomainName = HrJobDomain::all();
+        $domainName = HrJobDomain::all();
         $jobName = Job::all();
 
         return view('hr.employees.fte-handler')->with([
-            'DomainName' => $DomainName,
+            'domainName' => $domainName,
             'employees' => $employees,
             'jobName' => $jobName
         ]);
-    }
-
-    public function store(Request $request)
-    {
-        $jobrequisition = $request->validate([
-            'domain' => 'required|integer',
-            'job' => 'required|integer',
-        ]);
-
-        JobRequisition::create([
-            'domain_id' => $jobrequisition['domain'],
-            'job_id' => $jobrequisition['job']
-        ]);
-
-        $jobHiring = null;
-        Mail::send(new SendHiringMail($jobHiring));
-
-        return redirect()->back();
     }
 }
