+++ resources/views/emails/send-hiring.blade.php
@@ -3,6 +3,6 @@
 	<p>New job requisition has been added 
              </p>
    <a>check the requisition  </a>
-   <a href="http://portal.test/hr/requisition">click here</a>
+   <a href={{config('hr.url.requisition_url')}} >click here</a>
 </div><br>
 Thanks
