+++ resources/views/hr/employees/index.blade.php
@@ -26,53 +26,48 @@
             <th>Projects Count</th>
             <th>Current FTE</th>
         </tr>
-
         @foreach ($employees as $employee)
-        <tr>
-            <td>
-                <a href={{ route('employees.show', $employee->id) }}>{{ $employee->name }}</a>
-            </td>
-            <td>
-                @if ($employee->designation)
-                    {{ $employee->designation }}
+            <tr>
+                <td>
+                    <a href={{ route('employees.show', $employee->id) }}>{{ $employee->name }}</a>
+                </td>
+                <td>
+                    @if ($employee->designation)
+                        {{ $employee->designation }}
+                    @else
+                        -
+                    @endif
+                </td>
+                <td>
+                @if ($employee->joined_on)
+                    <span>{{$employee->joined_on->format('d M, Y') }}</span>
+                    <span style="font-size: 10px;">&nbsp;&#9679;&nbsp;</span>
+                    <span>{{$employee->employmentDuration }}</span>
                 @else
                     -
                 @endif
-            </td>
-            <td>
-            @if ($employee->joined_on)
-                <span>{{$employee->joined_on->format('d M, Y') }}</span>
-                <span style="font-size: 10px;">&nbsp;&#9679;&nbsp;</span>
-                <span>{{$employee->employmentDuration }}</span>
-            @else
-                -
-            @endif
-            </td>
-            <td>
-                @if($employee->user == null)
-                    0
-                @else
-                    {{count($employee->user->activeProjectTeamMembers)}}
-                @endif
-            </td>
-            <td>
-                @if ($employee->user == null)
-
-                  <span class="text-danger">{{ $employee->user ? $employee->user->fte :'NA' }}</span>
-
-                @elseif ($employee->user->fte > 1 && $employee->domain_id != null)
+                </td>
+                <td>
+                    @if($employee->user == null)
+                        0
+                    @else
+                        {{count($employee->user->activeProjectTeamMembers)}}
+                    @endif
+                </td>
+                <td>
+                    @if ($employee->user == null)
+                    <span class="text-danger">{{ $employee->user ? $employee->user->fte :'NA' }}</span>
+                    @elseif ($employee->user->fte > 1 && $employee->domain_id != null)
                     <a class="text-success" href={{ route('employees.alert',['domain_id' => $employee->domain_id]) }} style="text-decoration: none;">
                         {{ $employee->user->fte }} &nbsp;&nbsp;&nbsp;<span class="text-danger"><i class="fa fa-warning fa-lg"></i></span>
                     </a>
-                @else
-                <span class="text-danger">{{ $employee->user ? $employee->user->fte :'NA' }}</span>
-                @endif
-            </td>
+                    @else
+                    <span class="text-danger">{{ $employee->user ? $employee->user->fte :'NA' }}</span>
+                    @endif
+                </td>
             </tr>
-            
         @endforeach
     </table>
-
 </div>
 @endsection
