+++ resources/views/hr/requisition/menu.blade.php
@@ -3,6 +3,6 @@
         <a class="nav-item nav-link {{ Request::is('hr/requisition*') ? 'active' : '' }}" href="{{ route('requisition') }}"><i class="fa fa-users"></i>&nbsp;Pending</a>
     </li>
     <li class="nav-item">
-        <a class="nav-item nav-link {{ Request::is('hr/rquisition-complete*') ? 'active' : '' }}" href="{{ route('employees.complete') }}"><i class="fa fa-users"></i>&nbsp;Complete</a>
+        <a class="nav-item nav-link {{ Request::is('hr/complete*') ? 'active' : '' }}" href="{{ route('requisition.complete') }}"><i class="fa fa-users"></i>&nbsp;Complete</a>
     </li>
 </ul>
